# services/backend/blue-magma-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for blue magma auth and data",
        "title": "Bluemagma API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
...
        "/auth/google/exchange": {
            "post": {
                "description": "Exchanges Google OAuth authorization code for JWT tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Exchange Google OAuth code for tokens",
                "parameters": [
                    {
                        "description": "Code exchange request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.GoogleExchangeCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoogleOAuthCallbackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoogleOAuthCallbackResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoogleOAuthCallbackResponse"
                        }
                    }
                }
            }
        },
        "/auth/google/start": {
            "post": {
                "description": "Initiates Google OAuth flow for user login or account linking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Start Google OAuth flow",
                "parameters": [
                    {
                        "description": "OAuth start request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.GoogleStartOAuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoogleStartOAuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
...
        "handlers.GoogleExchangeCodeRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "\"login\" or \"link\"",
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "handlers.GoogleOAuthCallbackResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "handlers.GoogleStartOAuthRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "\"login\" or \"link\"",
                    "type": "string"
                },
                "return_url": {
                    "type": "string"
                }
            }
        },
        "handlers.GoogleStartOAuthResponse": {
            "type": "object",
            "properties": {
                "oauth_url": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and your JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}